requirements.txt

aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.0
django==3.1

django-bootstrap-modal-forms
django-crispy-forms
django-geojson
django-import-export
django-leaflet
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
filelock==3.0.12
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jdcal==1.4.1
jsonfield==3.1.0
MarkupPy==1.14
msgpack==0.6.2
numpy
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1

pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pytz==2019.3
PyYAML==5.3
reportlab==3.5.42
scipy
service-identity==18.1.0
six==1.14.0
sqlparse
tablib
Twisted
txaio==20.4.1
virtualenv==20.0.4
webencodings==0.5.1
win-unicode-console==0.5
xhtml2pdf
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
